# server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_request.BatchForceOfflineRequest:
    properties:
      reason:
        description: 下线原因
        type: string
      userIds:
        description: 用户 ID 列表
        items:
          type: string
        type: array
    required:
    - userIds
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.OnlineUserListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OnlineUserResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_ix-pay_ixpay-pro_internal_dto_base_response.OnlineUserResponse:
    properties:
      browser:
        type: string
      device:
        type: string
      lastActiveAt:
        type: string
      loginIp:
        type: string
      loginPlace:
        type: string
      loginTime:
        type: string
      nickname:
        type: string
      os:
        type: string
      sessionId:
        type: string
      userAgent:
        type: string
      userId:
        example: ""
        type: string
      username:
        type: string
    type: object
...
  github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8586
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于Gin和PostgreSQL的微信支付API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 微信支付API服务
  version: "1.0"
paths:
...
  /api/admin/online-user:
    get:
      consumes:
      - application/json
      description: 获取当前所有在线用户信息
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 在线用户列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OnlineUserListResponse'
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取在线用户列表
      tags:
      - 在线用户管理
  /api/admin/online-user/:user_id:
    delete:
      consumes:
      - application/json
      description: 强制指定用户下线（管理员权限）
      parameters:
      - description: 用户 ID
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: 下线原因
        in: body
        name: reason
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 强制下线成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不在线
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 强制用户下线
      tags:
      - 在线用户管理
    get:
      consumes:
      - application/json
      description: 根据用户 ID 获取在线用户详细信息
      parameters:
      - description: 用户 ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 在线用户详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_response.OnlineUserResponse'
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用户不在线
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取在线用户详情
      tags:
      - 在线用户管理
  /api/admin/online-user/batch:
    post:
      consumes:
      - application/json
      description: 批量强制指定用户下线（管理员权限）
      parameters:
      - description: 批量下线请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_dto_base_request.BatchForceOfflineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量强制下线成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 批量强制用户下线
      tags:
      - 在线用户管理
  /api/admin/online-user/count:
    get:
      consumes:
      - application/json
      description: 获取当前在线用户总数
      produces:
      - application/json
      responses:
        "200":
          description: 在线用户数量
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  format: int64
                  type: integer
                msg:
                  type: string
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取在线用户数量
      tags:
      - 在线用户管理
  /api/admin/online-user/online:
    get:
      consumes:
      - application/json
      description: 根据用户 ID 检查用户是否在线
      parameters:
      - description: 用户 ID
        format: int64
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 是否在线
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ix-pay_ixpay-pro_internal_utils_common_baseRes.Response'
            - properties:
                data:
                  type: boolean
                msg:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 检查用户是否在线
      tags:
      - 在线用户管理
...
securityDefinitions:
  BearerAuth:
    in: header [Bearer ]
    name: Authorization
    type: apiKey
swagger: "2.0"